+++ src/yass/preprocess/score.py
@@ -2,75 +2,77 @@
 """
 import os
 import numpy as np
+from ..geometry import order_channels_by_distance
 
 
-def getScore(spt, rot, n_channels, spike_size, n_features, neighbors,
-             wf_path, scale_to_save, n_batches, n_portions, buff,
-             batch_size):
+def get_score_pca(spike_index, rot, neighbors, geom, batch_size,
+             wf_path, scale_to_save):
     """PCA scoring
     """
+    # column ids for index matrix
+    SPIKE_TIME, MAIN_CHANNEL = 0, 1
+
+    window_size, n_features, n_channels = rot.shape
+    spike_size = int((window_size-1)/2)
+    n_spikes = spike_index.shape[0]
+
     wf_file = open(os.path.join(wf_path), 'rb')
-
-    batch_size = batch_size+2*buff
-
     flattenedLength = 2*batch_size*n_channels
 
-    neighchan = neighbors
-    C = n_channels
-    R = spike_size
+    nneigh = np.max(np.sum(neighbors, 0))
+    c_idx = np.ones((n_channels, nneigh), 'int32')*n_channels
+    for c in range(n_channels):
+        ch_idx, _ = order_channels_by_distance(c,
+                                               np.where(neighbors[c])[0],
+                                               geom)
+        c_idx[c,:ch_idx.shape[0]] = ch_idx
 
-    score = [None]*C
-    clr_idx = [None]*C
+    score = np.zeros((n_spikes, n_features, nneigh), 'float32')
 
-    nBatches = n_batches
-    nPortion = n_portions
+    counter_batch = 0
+    for i in np.unique(spike_index[:,2]):
+        spike_index_batch = spike_index[spike_index[:,2]==i]
+        n_spikes_batch = spike_index_batch.shape[0]
 
-    for i in range(0, nBatches):
-        if i <= nPortion:
+        wf_file.seek(flattenedLength*i)
+        wrec = wf_file.read(flattenedLength)
+        wrec = np.fromstring(wrec, dtype='int16')
+        wrec = np.reshape(wrec, (-1, n_channels))
+        wrec = wrec.astype('float32')/scale_to_save
+        wrec = np.concatenate((wrec, np.zeros((batch_size,1))), axis=1)
 
-            wf_file.seek(flattenedLength*i)
-            wrec = wf_file.read(flattenedLength)
+        nbuff = 50000
+        wf = np.zeros((nbuff, window_size, nneigh), 'float32')
+        count = 0
+        for j in range(n_spikes_batch):
+            t = spike_index_batch[j,SPIKE_TIME]
+            ch_idx = c_idx[spike_index_batch[j,MAIN_CHANNEL]]
+            wf[count] = wrec[(t-spike_size):(t+spike_size+1),ch_idx]
+            count += 1
 
-            wrec = np.fromstring(wrec, dtype='int16')
+            if (count == nbuff) or (j == n_spikes_batch -1):
+                # if we seek all spikes before reaching the buffer size,
+                # size of buffer becomes the number of leftover spikes
+                if j == n_spikes-1:
+                    nbuff = count
+                    wf = wf[:nbuff]
 
-            wrec = np.reshape(wrec, (-1, n_channels))
-            wrec = wrec.astype('float32')/scale_to_save
+                # calculate score and collect into variable 'score'
+                score_temp = np.zeros((wf.shape[0],n_features, nneigh))
+                for j in range(nneigh):
+                    if ch_idx[j] < n_channels:
+                        score_temp[:,:,j] = np.matmul(wf[:,:,j],rot[:,:,ch_idx[j]])
+                score[counter_batch:(counter_batch+nbuff)] = score_temp
 
-            for c in range(C):
-                ch_idx = np.where(neighchan[c])[0]
-                clr_idx_temp = np.where(spt[c][:, 1] == i)[0]
-                spt_c = spt[c][clr_idx_temp, 0]
-
-                if spt_c.shape[0] > 0:
-                    # get waveforms
-                    wf = np.zeros((spt_c.shape[0], 2*R+1, ch_idx.shape[0]))
-
-                    for j in range(spt_c.shape[0]):
-                        wf[j] = wrec[spt_c[j]+np.arange(-R, R+1)][:, ch_idx]
-
-                    score_temp = np.zeros(
-                        (wf.shape[0], n_features, wf.shape[2]))
-                    for j in range(ch_idx.shape[0]):
-                        score_temp[:, :, j] = np.matmul(
-                            wf[:, :, j], rot[:, :, ch_idx[j]])
-
-                    if i == 0:
-                        score[c] = score_temp
-                        clr_idx[c] = clr_idx_temp
-                    else:
-                        score[c] = np.concatenate((score[c], score_temp))
-                        clr_idx[c] = np.concatenate((clr_idx[c], clr_idx_temp))
-                else:
-                    if i == 0:
-                        score[c] = np.zeros((0, n_features, ch_idx.shape[0]))
-                        clr_idx[c] = np.zeros((0), 'int16')
-
+                # set counter back to zero
+                count = 0
+                counter_batch += nbuff
     wf_file.close()
 
-    return score, clr_idx
+    return score
 
 
-def getPcaSS(recordings, spike_times, spike_size, buff):
+def get_pca_suff_stat(recordings, spike_index, spike_size):
     """Get PCA SS matrix per recording channel
 
     Parameters
@@ -84,20 +86,24 @@ def getPcaSS(recordings, spike_times, spike_size, buff):
     buff:
         Buffer size
     """
+    # column ids for index matrix
+    SPIKE_TIME, MAIN_CHANNEL = 0, 1
+
     n_obs, n_channels = recordings.shape
     window_idx = range(-spike_size, spike_size+1)
     window_size = len(window_idx)
 
-    ss = np.zeros((window_size, window_size, n_channels))
-    spikes_per_channel = np.zeros(n_channels)
+    pca_suff_stat = np.zeros((window_size, window_size, n_channels))
+    spikes_per_channel = np.zeros(n_channels, 'int32')
 
     # iterate over every channel
     for c in range(n_channels):
         # get spikes times for the current channel
-        channel_spike_times = spike_times[c]
+        channel_spike_times = spike_index[
+            spike_index[:,MAIN_CHANNEL]==c, SPIKE_TIME]
         channel_spike_times = channel_spike_times[np.logical_and(
-                              (channel_spike_times > spike_size+buff),
-                              (channel_spike_times < n_obs-spike_size-buff))]
+                              (channel_spike_times > spike_size),
+                              (channel_spike_times < n_obs-spike_size-1))]
 
         channel_spikes = len(channel_spike_times)
 
@@ -109,22 +115,20 @@ def getPcaSS(recordings, spike_times, spike_size, buff):
             # fill in recording values for each spike time
             wf_temp[j, :] = recordings[channel_spike_times + window_idx[j], c]
 
-        ss[:, :, c] = np.matmul(wf_temp, wf_temp.transpose())
+        pca_suff_stat[:, :, c] = np.matmul(wf_temp, wf_temp.T)
 
         spikes_per_channel[c] = channel_spikes
 
-    spikes_per_channel = spikes_per_channel.astype('int')
-
-    return ss, spikes_per_channel
+    return pca_suff_stat, spikes_per_channel
 
 
-def getPCAProjection(ss, spikes_per_channel, n_features, neighbors):
+def get_pca_projection(ss, spikes_per_channel, n_features, neighbors):
     """Get PCA projection matrix per channel
 
     Parameters
     ----------
     ss: matrix
-        SS matrix as returned from getPcaSS
+        SS matrix as returned from get_pca_suff_stat
     spikes_per_channel: array
         Number of spikes per channel
     n_features: int