+++ CHANGELOG.rst
@@ -1,7 +1,22 @@
 Changelog
 =========
 
-0.1
----
+0.1.2dev
+--------
+
+* Updates to input/output structure
+* Adds new module for augmented spikes
+* Function names changes in score module
+* Simplified parameters for score module functions
+
+
+0.1.1 (2017-11-01)
+-------------------
+
+* Minor changes to setup.py for uploading to pypi
+
+
+0.1 (2017-11-01)
+-----------------
 
 * First release